debian: Fixup: Avoid fix dkms version
The autoconf-generated dist tarball version is now the same as the
package version. The command to fix the version for the dkms
source directory failed because it tried to move a directory onto
itself.

Avoid the fix. However, use the explicit package source version
to catch mismatch between upstream and package versions at build time
instead of dkms install time.

Change-Id: I30d264b173f4ecb0cd3db6116a6079be2a2b5265
Signed-off-by: Tzafrir Cohen <tzafrirc@nvidia.com>